+++ NEWS.md
@@ -1,5 +1,11 @@
 # What is New in *forstringr*
 
+## 0.1.1 (2023-04-10) 
+
+- Added `str_englue()`, an alias of `rlang::englue()`.
+
+- Added meaningful error messages for all the functions in the forstringr package.
+
 ## 0.0.1 (2022-08-29) 
 
 Added package logo, `community_data` (data containing whitespaces), `length_omit_na()`, and `str_extract_part()` functions, and modified unit tests.